评论+状态切换

+++ pages/posts/show/index.js
@@ -21,7 +21,13 @@ Page({
       content: '',
       to_user_id: 0,
       to_comment_id: 0
-    }
+    },
+
+    page: 1,
+    pageRows: 10,
+    isLastPage: 0,
+    list: [],
+    total_count: 0,
   },
 
   /**
@@ -65,6 +71,21 @@ Page({
     });
   },
 
+  ViewCommentListImage(e) {
+    console.log(e.currentTarget.dataset.key)
+    wx.previewImage({
+      urls: this.data.list[e.currentTarget.dataset.key].images,
+      current: e.currentTarget.dataset.url
+    });
+  },
+
+  ViewToCommentImage(e) {
+    wx.previewImage({
+      urls: this.data.list[e.currentTarget.dataset.key].toComment.images,
+      current: e.currentTarget.dataset.url
+    });
+  },
+
   bindToCreatePage: function(e) {
     wx.navigateTo({
       url: '/pages/posts/form/index?id=' + this.data.id,
@@ -77,8 +98,10 @@ Page({
       content: '确定要删除此帖子吗，无法恢复？',
       cancelText: '取消',
       confirmText: '确定',
-      success: res => {
-        this.deleteDo(this.data.id)
+      success: (res) => {
+        if (res.confirm) {
+          this.deleteDo(this.data.id)
+        }
       }
     })
   },
@@ -97,7 +120,9 @@ Page({
 
   showModal(e) {
     this.setData({
-      modalName: e.currentTarget.dataset.target
+      modalName: e.currentTarget.dataset.target,
+      'comment.to_user_id': e.currentTarget.dataset.uid || 0,
+      'comment.to_comment_id': e.currentTarget.dataset.id || 0,
     })
   },
   hideModal(e) {
@@ -155,36 +180,112 @@ Page({
       return 0;
     }
 
-    this.hideModal(e);
+    let that = this;
+    app.https.POST({
+      url: '/api/postComments/' + that.data.id,
+      params: that.data.comment,
+      success: function(res) {
+        that.setData({
+          comment: {
+            images: [],
+            content: '',
+            to_user_id: 0,
+            to_comment_id: 0
+          }
+        })
+        app.notice.showToast('评论成功');
+        that.hideModal();
+        that.onReady();
+      }
+    });
+  },
+
+  getList: function(isNext) {
+    let that = this;
+    app.https.GET({
+      url: '/api/postComments',
+      params: {
+        post_id: that.data.id,
+        page: that.data.page,
+        per_page: that.data.pageRows,
+        include: 'owner,toComment,toComment.owner',
+      },
+      success: function(res) {
+        for (let i = 0, len = res.data.data.length; i < len; i++) {
+          res.data.data[i].decode_content = res.data.data[i].content.split("\n")
+          if (res.data.data[i].toComment) {
+            res.data.data[i].toComment.decode_content = res.data.data[i].toComment.content.split("\n");
+          }
+        }
+        if (isNext) {
+          //获取下一页
+          that.setData({
+            total_count: res.data.meta.pagination.total,
+            list: that.data.list.concat(res.data.data),
+            isLastPage: res.data.meta.pagination.total_pages > res.data.meta.pagination.current_page ? 0 : 1,
+          });
+        } else {
+          that.setData({
+            total_count: res.data.meta.pagination.total,
+            list: res.data.data,
+            isLastPage: res.data.meta.pagination.total_pages > res.data.meta.pagination.current_page ? 0 : 1,
+          });
+        }
+      }
+    });
+  },
+
+  bindDeleteComment: function (e) {
+    let that = this;
+    wx.showModal({
+      title: '确定删除',
+      content: '确定要删除此评论吗，无法恢复？',
+      cancelText: '取消',
+      confirmText: '确定',
+      success: (res) => {
+        if (res.confirm) {
+          app.https.DELETE({
+            url: '/api/postComments/' + e.currentTarget.dataset.id,
+            success: function (res) {
+              app.notice.showToast('删除成功', 'success', () => {
+                that.onReady();
+              });
+            }
+          });
+        }
+      }
+    })
   },
 
   /**
    * 生命周期函数--监听页面初次渲染完成
    */
   onReady: function() {
-
+    this.getList(0);
   },
-
-  /**
-   * 生命周期函数--监听页面显示
-   */
-  onShow: function() {
-
-  },
-
 
   /**
    * 页面相关事件处理函数--监听用户下拉动作
    */
   onPullDownRefresh: function() {
-
+    this.getInfo(this.data.id);
+    this.getList(0);
+    this.setData({
+      user_info: app.globalData.userInfo
+    })
   },
 
   /**
    * 页面上拉触底事件的处理函数
    */
   onReachBottom: function() {
-
+    if (this.data.isLastPage) {
+      return false;
+    }
+    this.setData({
+      page: this.data.page + 1
+    });
+    this.getList(1);
   },
 
   /**
